+++ mars/mips/instructions/syscalls/SyscallInputDialogString.java
@@ -104,9 +104,9 @@ public class SyscallInputDialogString extends AbstractSyscall {
                             inputString.charAt(index));
                 }
                 if (inputString.length() < maxLength - 1) {
-                    Globals.memory.setByte(byteAddress + (int) Math.min(inputString.length(), maxLength - 2), '\n');  // newline at string end
+                    Globals.memory.setByte(byteAddress + Math.min(inputString.length(), maxLength - 2), '\n');  // newline at string end
                 }
-                Globals.memory.setByte(byteAddress + (int) Math.min((inputString.length() + 1), maxLength - 1), 0);  // null char to end string
+                Globals.memory.setByte(byteAddress + Math.min((inputString.length() + 1), maxLength - 1), 0);  // null char to end string
 
                 if (inputString.length() > maxLength - 1) {
                     //  length of the input string exceeded the specified maximum
