+++ mars/tools/KeyboardAndDisplaySimulator.java
@@ -867,7 +867,7 @@ public class KeyboardAndDisplaySimulator extends AbstractMarsToolAndApplication 
             public void stateChanged(ChangeEvent e) {
                 JSlider source = (JSlider) e.getSource();
                 if (!source.getValueIsAdjusting()) {
-                    delayLengthIndex = (int) source.getValue();
+                    delayLengthIndex = source.getValue();
                     transmitDelayInstructionCountLimit = generateDelay();
                 } else {
                     sliderLabel.setText(setLabel(source.getValue()));
